requirements.txt
fire==0.7.0
hydra-core==1.3.2
submitit==1.5.2
transformers==4.50.3
wandb==0.18.7
pillow==11.1.0
scikit-learn==1.6.1
rich==13.9.4
datasets==3.4.1
absl-py==2.2.2
peft==0.15.1
ml_collections==1.0.0
inflect==7.5.0
sentencepiece==0.2.0
timm==1.0.15
pydantic==1.10.19
accelerate==1.5.2
numpy==1.26.3
diffusers==0.31.0
clip==0.2.0